risk_engineering/02_research_reproducibility_template/requirements.txt
# Core libraries
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Experiment tracking
mlflow>=2.0.0
dvc>=3.0.0
wandb>=0.15.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0
pyyaml>=6.0

# ML frameworks
torch>=2.0.0
tensorflow>=2.15.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.17.0

# Documentation
jinja2>=3.1.0
tabulate>=0.9.0

# Utils
dill>=0.3.6
GitPython>=3.1.0
joblib>=1.3.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0